fix wrong strftime format
The format for strftime was very wrong which results in an issue where you can customize the values with an output filter.

I have no idea why you're changing the format for (and only for) publishedon, since it's very easy to do with output filter. But at least this change makes it possible to customize it again. :)